function/stdlib: Update tests for newer Go versions
Our function for parsing conventional timestamps strings currently leans
on the Go standard library "time" package's parser to help produce error
messages, but that does means that some of our error messages for invalid
timestamps end up being a little strange, and the details of that can
vary with changes to Go's time parser.

In future we should probably work to entirely divorce ourselves from Go's
implementation (we're already dealing with _most_ of the parsing ourselves
anyway) but for now we'll just update the expected result to match the
current versions of the upstream parser.